Just Kidding...
I'm going to work on a second solution for this. I'm not understanding something about the function of these loops. For some reason, it's returning the array values that I set at the very beginning of the function... effectively doing nothing. I am going to attempt to translate this into a for loop, which I think that I understand more. See chunky_monky2.j