Solution
This was straightforward. It felt a little bit silly when I got into the final testing cases. I was designing statements specifically for whatever the test case was showing should be returning true or false. Oh well! That's probably a normal aspect of this kind of work. I just never saw computers that way.